actually implement courtesy message lmao
i tested this by pasting the exact code in on_welcome and it seems to
work decently well. one edge case that i naturally found when doing that
was when on_welcome was called before the discord client finished
initializing, causing client.get_guild to return None and cause badness
in the next line. in practice this should not be a factor when
reconnecting, but that's why there's the silly timeout logic there. also
i swear i had an issue once where i got the guild correctly but then the
channel returned None??? i haven't been able to replicate that and it
makes no sense anyways so leaving that check out for now, but will keep
a note of it just in case.